requirements.txt
pyyaml>=6.0
pydantic>=2.0
pydantic[email]>=2.0
pytest>=7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
